// cosmowcl/src/app/shared/models/api/cosmoecm/documentiTask.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * firma.v1
 * API per la gestione della firma digitale
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Attivita } from './attivita.d';
import { Documento } from './documento.d';


export interface DocumentiTask { 
    documenti?: Array<Documento>;
    attivita?: Attivita;
}